LPS-196642 Inline SalesforceHttp into SalesforceObjectEntryManagerImpl
